MSFG-LoanTools-Pro/python-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.0.1
python-multipart==0.0.6
requests==2.31.0

# Advanced Tax Processor Dependencies
# OCR Engines
paddleocr>=2.6.0.3
paddlepaddle>=2.4.0

# Image Processing
opencv-python>=4.8.0
imgaug>=0.4.0

# Document Processing
PyPDF2>=3.0.0
pikepdf>=5.0.0,<9.0.0
pdfminer.six>=20221105
pdf2image>=1.16.3

# Table Extraction
camelot-py>=0.11.0
tabula-py>=2.7.0

# NLP and Validation
spacy>=3.6.0
durable-rules>=2.0.28
pydantic>=2.0.0

# Development and Testing
pytest>=7.3.1
black>=23.3.0
numpy==1.24.3
opencv-python==4.8.1.78


